Move mappers to `xsbti.compile.analysis`
The right place for mappers is in the Java universe, not `sbt.inc`. The
whole Zinc API is migrating away from placing public APIs in `sbt.inc`.
