+++ internal/zinc-persist/src/main/java/xsbti/compile/analysis/GenericMapper.java
@@ -5,10 +5,9 @@
  * This software is released under the terms written in LICENSE.
  */
 
-package sbt.inc;
+package xsbti.compile.analysis;
 
 import xsbti.compile.MiniSetup;
-import xsbti.compile.analysis.Stamp;
 
 import java.io.File;
