+++ internal/zinc-persist/src/main/java/xsbti/compile/analysis/ReadWriteMappers.java
@@ -5,7 +5,7 @@
  * This software is released under the terms written in LICENSE.
  */
 
-package sbt.inc;
+package xsbti.compile.analysis;
 
 import java.nio.file.Path;
