+++ internal/zinc-persist/src/main/java/xsbti/compile/analysis/WriteMapper.java
@@ -5,30 +5,29 @@
  * This software is released under the terms written in LICENSE.
  */
 
-package sbt.inc;
+package xsbti.compile.analysis;
 
-import sbt.internal.inc.mappers.RelativeReadMapper;
+import sbt.internal.inc.mappers.RelativeWriteMapper;
 import xsbti.compile.MiniSetup;
-import xsbti.compile.analysis.Stamp;
 
 import java.io.File;
 import java.nio.file.Path;
 
 /**
- * Defines a reader-only mapper interface that is used by Zinc after reading
- * the analysis from an instance of an {@link xsbti.compile.AnalysisStore}.
+ * Defines a writer-only mapper interface that is used by Zinc before writing
+ * the contents of the analysis files to the underlying analysis store.
  *
  * This interface is useful to make the analysis file machine-independent and
  * allow third parties to distribute them around.
  */
-public interface ReadMapper extends GenericMapper {
+public interface WriteMapper extends GenericMapper {
 
     /**
-     * Defines a mapper that reads from a machine-independent analysis file.
+     * Defines a mapper that writes a machine-independent analysis file.
      *
-     * An analysis file is machine independent if all the paths are relative to a path
-     * and no information about the machine that produced it is stored -- only
-     * information about the structure of the project from a given build tool is persisted.
+     * An analysis file is machine independent if all the paths are relative and no
+     * information about the machine that produced it is stored -- only information
+     * about the structure of the project from a given build tool is persisted.
      *
      * The mapper makes sure that the analysis file looks like if it was generated by
      * the local machine it's executed on.
@@ -44,24 +43,23 @@ public interface ReadMapper extends GenericMapper {
      * Build tools that need more careful handling of paths should create their own read and
      * write mappers.
      *
-     *
-     * @param projectRootPath The path on which we want to "mount" all the relative paths in analysis.
-     * @return A read mapper to pass in to {@link sbt.internal.inc.FileAnalysisStore}.
+     * @param projectRootPath The path on which the analysis file was generated and has to be made relative.
+     * @return A write mapper to pass in to {@link sbt.internal.inc.FileAnalysisStore}.
      */
-    static ReadMapper getMachineIndependentMapper(Path projectRootPath) {
-        return new RelativeReadMapper(projectRootPath);
+    static WriteMapper getMachineIndependentMapper(Path projectRootPath) {
+        return new RelativeWriteMapper(projectRootPath);
     }
 
     /**
-     * Defines an no-op read mapper.
+     * Defines an no-op write mapper.
      *
      * This is useful when users are not interested in distributing the analysis files
      * and need to pass a read mapper to {@link sbt.internal.inc.FileAnalysisStore}.
      *
      * @return A no-op read mapper.
      */
-    static ReadMapper getEmptyMapper() {
-        return new ReadMapper() {
+    static WriteMapper getEmptyMapper() {
+        return new WriteMapper() {
             @Override
             public File mapSourceFile(File sourceFile) {
                 return sourceFile;
